Add GitLab access token authentication to leaderboard.
This makes it so that netids are determined based on a call to the
GitLab API with a GitLab API Access Token. Students will be instructed
to set this API key in their repositories as a secret environment
variable, along with another variable for their desired competition
alias, which will now default to "Anonymous" if unspecified.